Refactor: Convert WishlistScreen to LazyColumn for better performance
This commit refactors the `WishlistScreen` to use `LazyColumn` for displaying the list of games. This improves performance, especially for longer wishlists, by only composing and laying out items that are currently visible on screen.

Previously, the screen used a standard `Column` with `forEach`, which could lead to performance issues with a large number of items.

The `modifier` for the main `Column` was also changed from `verticalScroll(rememberScrollState())` to `fillMaxSize()` for the `LazyColumn` to correctly handle scrolling. Other minor formatting changes were made for consistency.

Signed-off-by: Jchillah <jchillah@gmail.com>